requirements.txt
click==8.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyMySQL==1.0.2
six==1.16.0
Werkzeug==2.0.1
gunicorn==20.1.0
PyYAML==6.0
Flask==2.0.1
Flask-Login==0.6.2
cryptography==38.0.3
oauthlib==3.2.1
requests==2.28.1
boto3==1.26.3
pyOpenSSL==22.0.0
flask-mysqldb==1.0.1
Flask-SQLAlchemy==2.0.0
flask_marshmallow==0.14.0
smartystreets_python_sdk==4.11.9
#rds_db==